refactor: CQRS split - extract queries from repository, remove .interface.ts, use Luxon
- Extract ListerLogsQuery and ObtenirStatistiquesLogsQuery (inject prisma directly)
- Remove ListerLogsUseCase and ObtenirStatistiquesLogsUseCase (were just pass-throughs)
- Rename ApplicationLogRepository.interface.ts -> ApplicationLogRepository.ts (domain convention)
- Simplify ApplicationLogSQLRepository to only keep purger()
- PurgerLogsUseCase: use DateTime.now().minus({ days: 7 }) via Luxon
