+++ src/server/application-log/domain/ApplicationLogRepository.ts
@@ -1,0 +1,3 @@
+export interface ApplicationLogRepository {
+  purger(anterieurA: Date): Promise<number>;
+}
